+++ src/radical/pilot/__init__.py
@@ -37,9 +37,13 @@ from .compute_pilot_description import ComputePilotDescription
 from .resource_config    import ResourceConfig
 from .staging_directives import COPY, LINK, MOVE, TRANSFER, SKIP_FAILED, CREATE_PARENTS
 
-from .utils import version, version_detail, version_branch
-from .utils import sdist_name, sdist_path
-from .utils import logger
+
+from .utils  import version, version_detail, version_branch
+from .utils  import sdist_name, sdist_path
+from .utils  import logger
+
+# make worker submodule available
+import worker
 
 # ------------------------------------------------------------------------------
