+++ src/radical/pilot/agent/radical-pilot-agent-multicore.py
@@ -6660,131 +6660,6 @@ class AgentStagingOutputComponent(rpu.Component):
 
 # ==============================================================================
 #
-class AgentHeartbeatWorker(rpu.Worker):
-    """
-    The HeartbeatMonitor watches the command queue for heartbeat updates (and
-    other commands).
-    """
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg):
-
-        rpu.Worker.__init__(self, 'AgentHeartbeatWorker', cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def create(cls, cfg):
-
-        return cls(cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def initialize_child(self):
-
-        self._session_id    = self._cfg['session_id']
-        self._mongodb_url   = self._cfg['mongodb_url']
-
-        self.declare_idle_cb(self.idle_cb, self._cfg.get('heartbeat_interval'))
-
-        # all components use the command channel for control messages
-        self.declare_publisher ('command', rp.AGENT_COMMAND_PUBSUB)
-
-        self._pilot_id      = self._cfg['pilot_id']
-        self._session_id    = self._cfg['session_id']
-        self._runtime       = self._cfg['runtime']
-        self._starttime     = time.time()
-
-        # set up db connection
-        _, mongo_db, _, _, _  = ru.mongodb_connect(self._cfg['mongodb_url'])
-
-        self._p  = mongo_db["%s.p"  % self._session_id]
-        self._cu = mongo_db["%s.cu" % self._session_id]
-
-        # communicate successful startup
-        self.publish('command', {'cmd' : 'alive',
-                                 'arg' : self.cname})
-
-
-    # --------------------------------------------------------------------------
-    #
-    def finalize_child(self):
-
-        # communicate finalization
-        self.publish('command', {'cmd' : 'final',
-                                 'arg' : self.cname})
-
-
-    # --------------------------------------------------------------------------
-    #
-    def idle_cb(self):
-
-        try:
-            self._prof.prof('heartbeat', msg='Listen! Listen! Listen to the heartbeat!', uid=self._pilot_id)
-            self._check_commands()
-            self._check_state   ()
-            return True
-
-        except Exception as e:
-            self._log.exception('heartbeat died - cancel')
-            self.publish('command', {'cmd' : 'shutdown',
-                                     'arg' : 'exception'})
-
-    # --------------------------------------------------------------------------
-    #
-    def _check_commands(self):
-
-        # Check if there's a command waiting
-        retdoc = self._p.find_and_modify(
-                    query  = {"_id"  : self._pilot_id},
-                    update = {"$set" : {COMMAND_FIELD: []}}, # Wipe content of array
-                    fields = [COMMAND_FIELD]
-                    )
-
-        if not retdoc:
-            return
-
-        for command in retdoc[COMMAND_FIELD]:
-
-            cmd = command[COMMAND_TYPE]
-            arg = command[COMMAND_ARG]
-
-            self._prof.prof('ingest_cmd', msg="mongodb to HeartbeatMonitor (%s : %s)" % (cmd, arg), uid=self._pilot_id)
-
-            if cmd == COMMAND_CANCEL_PILOT:
-                self._log.info('cancel pilot cmd')
-                self.publish('command', {'cmd' : 'shutdown',
-                                         'arg' : 'cancel'})
-
-            elif cmd == COMMAND_CANCEL_COMPUTE_UNIT:
-                self._log.info('cancel unit cmd')
-                self.publish('command', {'cmd' : 'cancel_unit',
-                                         'arg' : command})
-
-            elif cmd == COMMAND_KEEP_ALIVE:
-                self._log.info('keepalive pilot cmd')
-                self.publish('command', {'cmd' : 'heartbeat',
-                                         'arg' : 'keepalive'})
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _check_state(self):
-
-        # Make sure that we haven't exceeded the agent runtime. if
-        # we have, terminate.
-        if time.time() >= self._starttime + (int(self._runtime) * 60):
-            self._log.info("Agent has reached runtime limit of %s seconds.", self._runtime*60)
-            self.publish('command', {'cmd' : 'shutdown',
-                                     'arg' : 'timeout'})
-
-
-
-# ==============================================================================
-#
 class AgentWorker(rpu.Worker):
 
     # --------------------------------------------------------------------------
@@ -7246,7 +7121,7 @@ class AgentWorker(rpu.Worker):
         if self.agent_name == 'agent_0':
             wmap = {
                 rp.AGENT_UPDATE_WORKER    : AgentUpdateWorker,
-                rp.AGENT_HEARTBEAT_WORKER : AgentHeartbeatWorker
+                rp.AGENT_HEARTBEAT_WORKER : rp.worker.Heartbeat
                 }
             for wname in wmap:
                 self._log.info('create worker %s', wname)
